// dist/rei-dot-com/types/foundations/cdr-color-action.d.ts
export interface CdrColorActionTokens {
  /**
   * @value #143528
   * @cssvar --cdr-color-action-border-brand
   */
  readonly CdrColorActionBorderBrand: string;
  /**
   * @value #d7d4ce
   * @cssvar --cdr-color-action-border-neutral-faint
   */
  readonly CdrColorActionBorderNeutralFaint: string;
  /**
   * @value #d44703
   * @cssvar --cdr-color-action-border-sale
   */
  readonly CdrColorActionBorderSale: string;
  /**
   * @value #5889d6
   * @cssvar --cdr-color-action-border-trigger-faint
   */
  readonly CdrColorActionBorderTriggerFaint: string;
  /**
   * @value #f7f7f5
   * @cssvar --cdr-color-action-icon-neutral-faint
   */
  readonly CdrColorActionIconNeutralFaint: string;
  /**
   * @value #d7d4ce
   * @cssvar --cdr-color-action-icon-neutral-subtle
   */
  readonly CdrColorActionIconNeutralSubtle: string;
  /**
   * @value #ffffff
   * @cssvar --cdr-color-action-icon-neutral-trace
   */
  readonly CdrColorActionIconNeutralTrace: string;
  /**
   * @value #d44703
   * @cssvar --cdr-color-action-icon-sale
   */
  readonly CdrColorActionIconSale: string;
  /**
   * @value #bfddca
   * @cssvar --cdr-color-action-surface-brand-faint
   */
  readonly CdrColorActionSurfaceBrandFaint: string;
  /**
   * @value #f7f7f5
   * @cssvar --cdr-color-action-surface-neutral-faint
   */
  readonly CdrColorActionSurfaceNeutralFaint: string;
  /**
   * @value #f1f0ed
   * @cssvar --cdr-color-action-surface-neutral-subtle
   */
  readonly CdrColorActionSurfaceNeutralSubtle: string;
  /**
   * @value #fcfcfb
   * @cssvar --cdr-color-action-surface-neutral-trace
   */
  readonly CdrColorActionSurfaceNeutralTrace: string;
  /**
   * @value #ffe9e0
   * @cssvar --cdr-color-action-surface-sale-faint
   */
  readonly CdrColorActionSurfaceSaleFaint: string;
  /**
   * @value #143528
   * @cssvar --cdr-color-action-text-brand
   */
  readonly CdrColorActionTextBrand: string;
  /**
   * @value #f7f7f5
   * @cssvar --cdr-color-action-text-neutral-faint
   */
  readonly CdrColorActionTextNeutralFaint: string;
  /**
   * @value #cecac3
   * @cssvar --cdr-color-action-text-neutral-subtle
   */
  readonly CdrColorActionTextNeutralSubtle: string;
  /**
   * @value #ffffff
   * @cssvar --cdr-color-action-text-neutral-trace
   */
  readonly CdrColorActionTextNeutralTrace: string;
  /**
   * @value #d44703
   * @cssvar --cdr-color-action-text-sale
   */
  readonly CdrColorActionTextSale: string;
  /**
   * @value #5889d6
   * @cssvar --cdr-color-action-text-trigger-faint
   */
  readonly CdrColorActionTextTriggerFaint: string;
}

export declare const CdrColorAction: CdrColorActionTokens;
export default CdrColorAction;